refactor(tests): split core_contract.rs into per-concern integration test files
Split 520-line monolithic core_contract.rs into focused test files:
- fasta_parsing.rs, fasta_validation.rs, tokenizer_contract.rs
- model_input_build.rs, fixture_contract.rs, api_parity.rs

QA: cargo test -p biors-core passes (26 tests)
